+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 6.0.2
 - Fix error in c-select when keypress event has no key.
+- Accept problem details error messages in API Callers.
 
 # 6.0.1
 - `createAspNetCoreHmrPlugin` now displays NPM package mismatches with the standard vite overlay instead of a custom overlay.